Initial commit
Adds node, graph, and connection data structure. Allows for adding/removing nodes and restructuring connections appropriately on removal. Also allows for CSV to be read in and parsed using PapaCSV parser.
Allows traversing the graph from a start node that will find the shortest path to all other nodes, then return a sorted list based on the total of the edges to get to that node.